Looks like I got the basics working now. In addition to the previously mentioned two annoyances, it also turns out that whiel HLSL matrices are row major, you don't have to switch the multiplication arguments because it does that for you. This explains why things looked so fucked up all the time even while I randomly tried changing things, it was doing not one but two transformations the wrong way around.
Since this appears to be working, I cleaned up all the random crap and put in some nice comments about what the hell is going on now. I also replaced the tangent space with the object space in the vertex shader so there are less transformations to be done in the surface shader, since we do everything in object space. It also makes it a little less cluttered, and the tangent space provided by Unity was nonorthogonal anyways.