And the suddenly, progress! The surface shader is pretty much done. Also added textures for the two remaining layers. The dirt layer took a dirt nap, turns out even 15 textures is too many (Unity uses two itself somewhere). I never really knew what to do with that layer anyways.
Layers are now blended together according to all the heights and stuff, and the blend weights now depend on height and slope. I set up a basic planet, I'm not a super big fan of the current look but for now it'll do. Colors are too dark and I'm not sure I like the fringey edges. I think the rock texture needs to have its "canyons" removed.

The surface shader is fully converted to scaling off the surface radius (or the average surface height, or the average water height). Now the surface script should be converted similarly as well. Then the radius controls everything, and the planet can be generated at any size.