Fixed shadowing in my scene. Trick was to increase the shadow distance, and turn off shadow cascades since that is buggy at large distances. Antony's scene is way smaller (smaller radius) so shadowing probably works there already. Also turned off shadows related to the atmosphere. The water does cast shadows, I think that is the best solution. I'm not interested in doing it "right" for that, I just want to use the standard Unity stuff and that's it. That also allows the end-user the greatest freedom.
I thought I had to add a lightmap and everything, lol. Still have to do that to support baked lighting though... so I included the unwrapped I did of the planet sphere I made. But it requires a lot of fuzz in the code, so I'm keeping the simple one for now.