oh holy hell it worked out of the box
only problem is the water, it doesn't write to the depth buffer so the atmosphere pretends it's underwater too. If I don't use the depth buffer for water, but its level, I should be able to fix this correctly completely (with even the atmosphere on the other side of the water showing up! This will need an approximation of the water's total/lowest transparency I guess... Hm.