Fixed atmosphere height controls being the wrong way around.
Disabled everything about the surface shader we don't need in the atmosphere shader. Not much is left, and one could argue we might as well create a vertex/fragment shader instead, but I'd like to keep everything as far away from having to do everything lighting related by hand as I can. Also, this way it's still possible for the atmosphere shader to be extended in the future.

I also looked up how exactly the Unity lighting system does the whole mutiple light mumbojumbo, and I think the best option is to disable the vertex lighting (i.e. spherical harmonics/light probes and non-important point and directional lights), dealing only with the per-pixel lights. These can either be directional lights or point lights. That way we always have a light direction (either a direction if its W component is 0 or a position if its W component is 1) so we can do our faux atmospheric shadowing. It does mean smaller lights do not light the atmosphere, but I think that might actually be better given how everything is implemented.